MoviePage refactored, now having three different sections for clearer understanding of the code
